Added unit test to verify domain binding in the config document
