--- sdk/src/test/java/com/microsoft/did/sdk/DnsBindingServiceTest.kt
@@ -1,58 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved
-
-package com.microsoft.did.sdk
-
-import com.microsoft.did.sdk.credential.service.models.serviceResponses.DnsBindingResponse
-import com.microsoft.did.sdk.crypto.protocols.jose.jws.JwsToken
-import com.microsoft.did.sdk.identifier.resolvers.Resolver
-import com.microsoft.did.sdk.util.serializer.Serializer
-import io.mockk.mockk
-import kotlinx.coroutines.runBlocking
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.Test
-
-class DnsBindingServiceTest {
-
-    private val resolver: Resolver = mockk()
-    private val serializer = Serializer()
-//    private val dnsBindingService = DnsBindingService(mockk(relaxed = true), resolver, serializer)
-
-/*    @Test
-    fun `test deserialize config doc jwt`() {
-        val docJwt = "eyJhbGciOiJFUzI1NksiLCJraWQiOiJkaWQ6aW9uOkVpRElqcmFxaTI2T0ptRzN6TTc0TGYwRzZjRWpQWXNDb1VLcDFXT1gzT0FFd2c_LWlvbi1pbml0aWFsLXN0YXRlPWV5SmtaV3gwWVY5b1lYTm9Jam9pUldsRVVFbGZTRUpNVFVoWGRqWk9WMkk1Wm1nMFUwRXpkVGx3ZFVSYVdsUktUelpQZUVKc1RXa3RjM0JMVVNJc0luSmxZMjkyWlhKNVgyTnZiVzFwZEcxbGJuUWlPaUpGYVVKUVdUVTBhR3hJWkRodmRHWlBTMXBsTW13MFltaDNiVmc1TkdkdFRuTXdRM1Y1Ym5sNWFWZDVZbkJSSW4wLmV5SjFjR1JoZEdWZlkyOXRiV2wwYldWdWRDSTZJa1ZwUkdVM2QycHhRVUpoWW1OcWJGWm1OMjVqUkhsSWJHaFJaVE5oTW1kVFFrWnNZbWxKWTJwb1NVTnRPV2NpTENKd1lYUmphR1Z6SWpwYmV5SmhZM1JwYjI0aU9pSnlaWEJzWVdObElpd2laRzlqZFcxbGJuUWlPbnNpY0hWaWJHbGpYMnRsZVhNaU9sdDdJbWxrSWpvaWMybG5YMk5tTTJFek1HSm1JaXdpZEhsd1pTSTZJa1ZqWkhOaFUyVmpjREkxTm1zeFZtVnlhV1pwWTJGMGFXOXVTMlY1TWpBeE9TSXNJbXAzYXlJNmV5SnJkSGtpT2lKRlF5SXNJbU55ZGlJNkluTmxZM0F5TlRack1TSXNJbmdpT2lKUGFWaGFNelJJU0ZGdlRUSkZhelJQVTBSeGVqVlJWamQzUlhoSlVHdHNPV0phV0hsR1QxSm5kbE5aSWl3aWVTSTZJa2QwU0hOSmFGTm1lamhhY25KRVFXOVFRMm8xVDBoYVgyZG9abVF6V0dwV1RXcG9URzl6Tmxaa2FUZ2lmU3dpY0hWeWNHOXpaU0k2V3lKaGRYUm9JaXdpWjJWdVpYSmhiQ0pkZlYxOWZWMTkjc2lnX2NmM2EzMGJmIn0.eyJAY29udGV4dCI6Imh0dHBzOi8vaWRlbnRpdHkuZm91bmRhdGlvbi8ud2VsbC1rbm93bi9jb250ZXh0cy9kaWQtY29uZmlndXJhdGlvbi12MC4wLmpzb25sZCIsImxpbmtlZF9kaWRzIjpbeyJzdWIiOiJkaWQ6aW9uOkVpRElqcmFxaTI2T0ptRzN6TTc0TGYwRzZjRWpQWXNDb1VLcDFXT1gzT0FFd2c_LWlvbi1pbml0aWFsLXN0YXRlPWV5SmtaV3gwWVY5b1lYTm9Jam9pUldsRVVFbGZTRUpNVFVoWGRqWk9WMkk1Wm1nMFUwRXpkVGx3ZFVSYVdsUktUelpQZUVKc1RXa3RjM0JMVVNJc0luSmxZMjkyWlhKNVgyTnZiVzFwZEcxbGJuUWlPaUpGYVVKUVdUVTBhR3hJWkRodmRHWlBTMXBsTW13MFltaDNiVmc1TkdkdFRuTXdRM1Y1Ym5sNWFWZDVZbkJSSW4wLmV5SjFjR1JoZEdWZlkyOXRiV2wwYldWdWRDSTZJa1ZwUkdVM2QycHhRVUpoWW1OcWJGWm1OMjVqUkhsSWJHaFJaVE5oTW1kVFFrWnNZbWxKWTJwb1NVTnRPV2NpTENKd1lYUmphR1Z6SWpwYmV5SmhZM1JwYjI0aU9pSnlaWEJzWVdObElpd2laRzlqZFcxbGJuUWlPbnNpY0hWaWJHbGpYMnRsZVhNaU9sdDdJbWxrSWpvaWMybG5YMk5tTTJFek1HSm1JaXdpZEhsd1pTSTZJa1ZqWkhOaFUyVmpjREkxTm1zeFZtVnlhV1pwWTJGMGFXOXVTMlY1TWpBeE9TSXNJbXAzYXlJNmV5SnJkSGtpT2lKRlF5SXNJbU55ZGlJNkluTmxZM0F5TlRack1TSXNJbmdpT2lKUGFWaGFNelJJU0ZGdlRUSkZhelJQVTBSeGVqVlJWamQzUlhoSlVHdHNPV0phV0hsR1QxSm5kbE5aSWl3aWVTSTZJa2QwU0hOSmFGTm1lamhhY25KRVFXOVFRMm8xVDBoYVgyZG9abVF6V0dwV1RXcG9URzl6Tmxaa2FUZ2lmU3dpY0hWeWNHOXpaU0k2V3lKaGRYUm9JaXdpWjJWdVpYSmhiQ0pkZlYxOWZWMTkiLCJpc3MiOiJkaWQ6aW9uOkVpRElqcmFxaTI2T0ptRzN6TTc0TGYwRzZjRWpQWXNDb1VLcDFXT1gzT0FFd2c_LWlvbi1pbml0aWFsLXN0YXRlPWV5SmtaV3gwWVY5b1lYTm9Jam9pUldsRVVFbGZTRUpNVFVoWGRqWk9WMkk1Wm1nMFUwRXpkVGx3ZFVSYVdsUktUelpQZUVKc1RXa3RjM0JMVVNJc0luSmxZMjkyWlhKNVgyTnZiVzFwZEcxbGJuUWlPaUpGYVVKUVdUVTBhR3hJWkRodmRHWlBTMXBsTW13MFltaDNiVmc1TkdkdFRuTXdRM1Y1Ym5sNWFWZDVZbkJSSW4wLmV5SjFjR1JoZEdWZlkyOXRiV2wwYldWdWRDSTZJa1ZwUkdVM2QycHhRVUpoWW1OcWJGWm1OMjVqUkhsSWJHaFJaVE5oTW1kVFFrWnNZbWxKWTJwb1NVTnRPV2NpTENKd1lYUmphR1Z6SWpwYmV5SmhZM1JwYjI0aU9pSnlaWEJzWVdObElpd2laRzlqZFcxbGJuUWlPbnNpY0hWaWJHbGpYMnRsZVhNaU9sdDdJbWxrSWpvaWMybG5YMk5tTTJFek1HSm1JaXdpZEhsd1pTSTZJa1ZqWkhOaFUyVmpjREkxTm1zeFZtVnlhV1pwWTJGMGFXOXVTMlY1TWpBeE9TSXNJbXAzYXlJNmV5SnJkSGtpT2lKRlF5SXNJbU55ZGlJNkluTmxZM0F5TlRack1TSXNJbmdpT2lKUGFWaGFNelJJU0ZGdlRUSkZhelJQVTBSeGVqVlJWamQzUlhoSlVHdHNPV0phV0hsR1QxSm5kbE5aSWl3aWVTSTZJa2QwU0hOSmFGTm1lamhhY25KRVFXOVFRMm8xVDBoYVgyZG9abVF6V0dwV1RXcG9URzl6Tmxaa2FUZ2lmU3dpY0hWeWNHOXpaU0k2V3lKaGRYUm9JaXdpWjJWdVpYSmhiQ0pkZlYxOWZWMTkiLCJuYmYiOjE2MDIyNzM2NDUsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIiwiaHR0cHM6Ly9pZGVudGl0eS5mb3VuZGF0aW9uLy53ZWxsLWtub3duL2NvbnRleHRzL2RpZC1jb25maWd1cmF0aW9uLXYwLjAuanNvbmxkIl0sImlzc3VlciI6ImRpZDppb246RWlESWpyYXFpMjZPSm1HM3pNNzRMZjBHNmNFalBZc0NvVUtwMVdPWDNPQUV3Zz8taW9uLWluaXRpYWwtc3RhdGU9ZXlKa1pXeDBZVjlvWVhOb0lqb2lSV2xFVUVsZlNFSk1UVWhYZGpaT1YySTVabWcwVTBFemRUbHdkVVJhV2xSS1R6WlBlRUpzVFdrdGMzQkxVU0lzSW5KbFkyOTJaWEo1WDJOdmJXMXBkRzFsYm5RaU9pSkZhVUpRV1RVMGFHeElaRGh2ZEdaUFMxcGxNbXcwWW1oM2JWZzVOR2R0VG5Nd1EzVjVibmw1YVZkNVluQlJJbjAuZXlKMWNHUmhkR1ZmWTI5dGJXbDBiV1Z1ZENJNklrVnBSR1UzZDJweFFVSmhZbU5xYkZabU4yNWpSSGxJYkdoUlpUTmhNbWRUUWtac1ltbEpZMnBvU1VOdE9XY2lMQ0p3WVhSamFHVnpJanBiZXlKaFkzUnBiMjRpT2lKeVpYQnNZV05sSWl3aVpHOWpkVzFsYm5RaU9uc2ljSFZpYkdsalgydGxlWE1pT2x0N0ltbGtJam9pYzJsblgyTm1NMkV6TUdKbUlpd2lkSGx3WlNJNklrVmpaSE5oVTJWamNESTFObXN4Vm1WeWFXWnBZMkYwYVc5dVMyVjVNakF4T1NJc0ltcDNheUk2ZXlKcmRIa2lPaUpGUXlJc0ltTnlkaUk2SW5ObFkzQXlOVFpyTVNJc0luZ2lPaUpQYVZoYU16UklTRkZ2VFRKRmF6UlBVMFJ4ZWpWUlZqZDNSWGhKVUd0c09XSmFXSGxHVDFKbmRsTlpJaXdpZVNJNklrZDBTSE5KYUZObWVqaGFjbkpFUVc5UVEybzFUMGhhWDJkb1ptUXpXR3BXVFdwb1RHOXpObFprYVRnaWZTd2ljSFZ5Y0c5elpTSTZXeUpoZFhSb0lpd2laMlZ1WlhKaGJDSmRmVjE5ZlYxOSIsImlzc3VhbmNlRGF0ZSI6IjIwMjAtMTAtMDlUMjA6MDA6NDUuMzExWiIsInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLCJEb21haW5MaW5rYWdlQ3JlZGVudGlhbCJdLCJjcmVkZW50aWFsU3ViamVjdCI6eyJpZCI6ImRpZDppb246RWlESWpyYXFpMjZPSm1HM3pNNzRMZjBHNmNFalBZc0NvVUtwMVdPWDNPQUV3Zz8taW9uLWluaXRpYWwtc3RhdGU9ZXlKa1pXeDBZVjlvWVhOb0lqb2lSV2xFVUVsZlNFSk1UVWhYZGpaT1YySTVabWcwVTBFemRUbHdkVVJhV2xSS1R6WlBlRUpzVFdrdGMzQkxVU0lzSW5KbFkyOTJaWEo1WDJOdmJXMXBkRzFsYm5RaU9pSkZhVUpRV1RVMGFHeElaRGh2ZEdaUFMxcGxNbXcwWW1oM2JWZzVOR2R0VG5Nd1EzVjVibmw1YVZkNVluQlJJbjAuZXlKMWNHUmhkR1ZmWTI5dGJXbDBiV1Z1ZENJNklrVnBSR1UzZDJweFFVSmhZbU5xYkZabU4yNWpSSGxJYkdoUlpUTmhNbWRUUWtac1ltbEpZMnBvU1VOdE9XY2lMQ0p3WVhSamFHVnpJanBiZXlKaFkzUnBiMjRpT2lKeVpYQnNZV05sSWl3aVpHOWpkVzFsYm5RaU9uc2ljSFZpYkdsalgydGxlWE1pT2x0N0ltbGtJam9pYzJsblgyTm1NMkV6TUdKbUlpd2lkSGx3WlNJNklrVmpaSE5oVTJWamNESTFObXN4Vm1WeWFXWnBZMkYwYVc5dVMyVjVNakF4T1NJc0ltcDNheUk2ZXlKcmRIa2lPaUpGUXlJc0ltTnlkaUk2SW5ObFkzQXlOVFpyTVNJc0luZ2lPaUpQYVZoYU16UklTRkZ2VFRKRmF6UlBVMFJ4ZWpWUlZqZDNSWGhKVUd0c09XSmFXSGxHVDFKbmRsTlpJaXdpZVNJNklrZDBTSE5KYUZObWVqaGFjbkpFUVc5UVEybzFUMGhhWDJkb1ptUXpXR3BXVFdwb1RHOXpObFprYVRnaWZTd2ljSFZ5Y0c5elpTSTZXeUpoZFhSb0lpd2laMlZ1WlhKaGJDSmRmVjE5ZlYxOSIsIm9yaWdpbiI6Ind3dy5nb29nbGUuY29tIn19fV19.MXxIIJHPTHQzu8UIR-39SezOnSU1hMfzHxWkUAoKDhO5BIG5f2awuX7ooRmuBpevm4VBPqsVVSA9aVwjqQ9nmA"
-        val domLinkCreds = dnsBindingService.deserializeConfigDocument(docJwt)
-        val expectedDid = "did:ion:EiDIjraqi26OJmG3zM74Lf0G6cEjPYsCoUKp1WOX3OAEwg?-ion-initial-state=eyJkZWx0YV9oYXNoIjoiRWlEUElfSEJMTUhXdjZOV2I5Zmg0U0EzdTlwdURaWlRKTzZPeEJsTWktc3BLUSIsInJlY292ZXJ5X2NvbW1pdG1lbnQiOiJFaUJQWTU0aGxIZDhvdGZPS1plMmw0Ymh3bVg5NGdtTnMwQ3V5bnl5aVd5YnBRIn0.eyJ1cGRhdGVfY29tbWl0bWVudCI6IkVpRGU3d2pxQUJhYmNqbFZmN25jRHlIbGhRZTNhMmdTQkZsYmlJY2poSUNtOWciLCJwYXRjaGVzIjpbeyJhY3Rpb24iOiJyZXBsYWNlIiwiZG9jdW1lbnQiOnsicHVibGljX2tleXMiOlt7ImlkIjoic2lnX2NmM2EzMGJmIiwidHlwZSI6IkVjZHNhU2VjcDI1NmsxVmVyaWZpY2F0aW9uS2V5MjAxOSIsImp3ayI6eyJrdHkiOiJFQyIsImNydiI6InNlY3AyNTZrMSIsIngiOiJPaVhaMzRISFFvTTJFazRPU0RxejVRVjd3RXhJUGtsOWJaWHlGT1JndlNZIiwieSI6Ikd0SHNJaFNmejhacnJEQW9QQ2o1T0haX2doZmQzWGpWTWpoTG9zNlZkaTgifSwicHVycG9zZSI6WyJhdXRoIiwiZ2VuZXJhbCJdfV19fV19"
-        assertThat(domLinkCreds.size).isGreaterThan(0)
-        assertThat(domLinkCreds.first().iss).isEqualTo(expectedDid)
-        assertThat(domLinkCreds.first().sub).isEqualTo(expectedDid)
-        assertThat(domLinkCreds.first().vc.issuer).isEqualTo(expectedDid)
-        assertThat(domLinkCreds.first().vc.credentialSubject.id).isEqualTo(expectedDid)
-    }*/
-
-/*    @Test
-    fun `test to get service endpoint from did`() {
-        val suppliedDid = "did:ion:EiAkfaAsQG7FageU5GQtIcVI73AphEyFydjRWVj-2Ww7dw?-ion-initial-state=eyJkZWx0YV9oYXNoIjoiRWlDQnZNU2RqTGRyZkNCZ1NjeXoyNFFkSmZnYXR5UGN1Y0dXeEJZaldvOEN5QSIsInJlY292ZXJ5X2NvbW1pdG1lbnQiOiJFaURsVWViZTVsalJLVkV6bUhabjNNTkpkOWdNcXV4VzJDb0N3bG1TSmxSVmpnIn0.eyJ1cGRhdGVfY29tbWl0bWVudCI6IkVpQUVnYTA5bUdwWWJVTjBiampCZGMxTjdGY2M0NHhtWTdVdkZqRDRZSjZRanciLCJwYXRjaGVzIjpbeyJhY3Rpb24iOiJyZXBsYWNlIiwiZG9jdW1lbnQiOnsicHVibGljX2tleXMiOlt7ImlkIjoic2lnbiIsInR5cGUiOiJFY2RzYVNlY3AyNTZrMVZlcmlmaWNhdGlvbktleTIwMTkiLCJqd2siOnsia3R5IjoiRUMiLCJ1c2UiOiJzaWciLCJhbGciOiJFUzI1NksiLCJjcnYiOiJzZWNwMjU2azEiLCJ4IjoiMWI5ZDJXWEx5Mm81V3JqVUtJQTRHRHR2QjcxQmlsdG5aRS03ZGQ5SzFSUSIsInkiOiJxVk1ZX19yOVpTWV9CeTZmSFFQN0R4c184c2k5bVNRelUtdENiWG50VW1VIn0sInB1cnBvc2UiOlsiYXV0aCIsImdlbmVyYWwiXX1dfX1dfQ"
-        runBlocking {
-            val serviceEndpoints = dnsBindingService.getDomainForRp(suppliedDid)
-            assertThat(serviceEndpoints).isEmpty()
-        }
-    }*/
-
-/*    @Test
-    fun `test to validate well-known config doc`() {
-        val docJwt =
-            """{
-    "@context": "https://identity.foundation/.well-known/contexts/did-configuration-v0.0.jsonld",
-    "linked_dids": [
-        "eyJhbGciOiJFUzI1NksiLCJraWQiOiJkaWQ6aW9uOkVpQjdKLWswZjNJbUctenlGVUZSbVY0Z2FPc2Rmankza2RkSl9kOU9PdjdzaEE_LWlvbi1pbml0aWFsLXN0YXRlPWV5SmtaV3gwWVY5b1lYTm9Jam9pUldsQ1psQlVXakZZYlRSNFRtTTVOblpRY25kdldUQTFia3hmTkRWWlVHMTViMVZuYjNsU2FrRkxRVTFrWnlJc0luSmxZMjkyWlhKNVgyTnZiVzFwZEcxbGJuUWlPaUpGYVVGRlRrMDBjRTVMWWtaNlUyWXhSazFZUkU1c1lUQlZObEZDVkcxclkxQkhPRTF3VEdOUlpGWlBUa3RuSW4wLmV5SjFjR1JoZEdWZlkyOXRiV2wwYldWdWRDSTZJa1ZwUWkxRFRUVTFTRjkyTjB4YVFXOUNXRWsyT1haRWNuQTFYMmQwWDJwdFF6SnVibGhDTFZVNVRsUnhWMmNpTENKd1lYUmphR1Z6SWpwYmV5SmhZM1JwYjI0aU9pSnlaWEJzWVdObElpd2laRzlqZFcxbGJuUWlPbnNpY0hWaWJHbGpYMnRsZVhNaU9sdDdJbWxrSWpvaWMybG5Yek0zTXpkbU1qSTJJaXdpZEhsd1pTSTZJa1ZqWkhOaFUyVmpjREkxTm1zeFZtVnlhV1pwWTJGMGFXOXVTMlY1TWpBeE9TSXNJbXAzYXlJNmV5SnJkSGtpT2lKRlF5SXNJbU55ZGlJNkluTmxZM0F5TlRack1TSXNJbmdpT2lKc2NrNVBZMTh5YjNSak0zWlVlRWx1WlRCbGNraFVVeTFHYkdGNE9VSnJRMkZXZWtKTFZGVmZaMkpuSWl3aWVTSTZJblZVTkZkNGJubFJZVnAzTFRSNlRsaEJkVE0wZGxscVNuWjNaRmh1WkRKNFdqZHhWbVZoYW1kMGMwVWlmU3dpY0hWeWNHOXpaU0k2V3lKaGRYUm9JaXdpWjJWdVpYSmhiQ0pkZlYxOWZWMTkjc2lnXzM3MzdmMjI2In0.eyJzdWIiOiJkaWQ6aW9uOkVpQjdKLWswZjNJbUctenlGVUZSbVY0Z2FPc2Rmankza2RkSl9kOU9PdjdzaEE_LWlvbi1pbml0aWFsLXN0YXRlPWV5SmtaV3gwWVY5b1lYTm9Jam9pUldsQ1psQlVXakZZYlRSNFRtTTVOblpRY25kdldUQTFia3hmTkRWWlVHMTViMVZuYjNsU2FrRkxRVTFrWnlJc0luSmxZMjkyWlhKNVgyTnZiVzFwZEcxbGJuUWlPaUpGYVVGRlRrMDBjRTVMWWtaNlUyWXhSazFZUkU1c1lUQlZObEZDVkcxclkxQkhPRTF3VEdOUlpGWlBUa3RuSW4wLmV5SjFjR1JoZEdWZlkyOXRiV2wwYldWdWRDSTZJa1ZwUWkxRFRUVTFTRjkyTjB4YVFXOUNXRWsyT1haRWNuQTFYMmQwWDJwdFF6SnVibGhDTFZVNVRsUnhWMmNpTENKd1lYUmphR1Z6SWpwYmV5SmhZM1JwYjI0aU9pSnlaWEJzWVdObElpd2laRzlqZFcxbGJuUWlPbnNpY0hWaWJHbGpYMnRsZVhNaU9sdDdJbWxrSWpvaWMybG5Yek0zTXpkbU1qSTJJaXdpZEhsd1pTSTZJa1ZqWkhOaFUyVmpjREkxTm1zeFZtVnlhV1pwWTJGMGFXOXVTMlY1TWpBeE9TSXNJbXAzYXlJNmV5SnJkSGtpT2lKRlF5SXNJbU55ZGlJNkluTmxZM0F5TlRack1TSXNJbmdpT2lKc2NrNVBZMTh5YjNSak0zWlVlRWx1WlRCbGNraFVVeTFHYkdGNE9VSnJRMkZXZWtKTFZGVmZaMkpuSWl3aWVTSTZJblZVTkZkNGJubFJZVnAzTFRSNlRsaEJkVE0wZGxscVNuWjNaRmh1WkRKNFdqZHhWbVZoYW1kMGMwVWlmU3dpY0hWeWNHOXpaU0k2V3lKaGRYUm9JaXdpWjJWdVpYSmhiQ0pkZlYxOWZWMTkiLCJpc3MiOiJkaWQ6aW9uOkVpQjdKLWswZjNJbUctenlGVUZSbVY0Z2FPc2Rmankza2RkSl9kOU9PdjdzaEE_LWlvbi1pbml0aWFsLXN0YXRlPWV5SmtaV3gwWVY5b1lYTm9Jam9pUldsQ1psQlVXakZZYlRSNFRtTTVOblpRY25kdldUQTFia3hmTkRWWlVHMTViMVZuYjNsU2FrRkxRVTFrWnlJc0luSmxZMjkyWlhKNVgyTnZiVzFwZEcxbGJuUWlPaUpGYVVGRlRrMDBjRTVMWWtaNlUyWXhSazFZUkU1c1lUQlZObEZDVkcxclkxQkhPRTF3VEdOUlpGWlBUa3RuSW4wLmV5SjFjR1JoZEdWZlkyOXRiV2wwYldWdWRDSTZJa1ZwUWkxRFRUVTFTRjkyTjB4YVFXOUNXRWsyT1haRWNuQTFYMmQwWDJwdFF6SnVibGhDTFZVNVRsUnhWMmNpTENKd1lYUmphR1Z6SWpwYmV5SmhZM1JwYjI0aU9pSnlaWEJzWVdObElpd2laRzlqZFcxbGJuUWlPbnNpY0hWaWJHbGpYMnRsZVhNaU9sdDdJbWxrSWpvaWMybG5Yek0zTXpkbU1qSTJJaXdpZEhsd1pTSTZJa1ZqWkhOaFUyVmpjREkxTm1zeFZtVnlhV1pwWTJGMGFXOXVTMlY1TWpBeE9TSXNJbXAzYXlJNmV5SnJkSGtpT2lKRlF5SXNJbU55ZGlJNkluTmxZM0F5TlRack1TSXNJbmdpT2lKc2NrNVBZMTh5YjNSak0zWlVlRWx1WlRCbGNraFVVeTFHYkdGNE9VSnJRMkZXZWtKTFZGVmZaMkpuSWl3aWVTSTZJblZVTkZkNGJubFJZVnAzTFRSNlRsaEJkVE0wZGxscVNuWjNaRmh1WkRKNFdqZHhWbVZoYW1kMGMwVWlmU3dpY0hWeWNHOXpaU0k2V3lKaGRYUm9JaXdpWjJWdVpYSmhiQ0pkZlYxOWZWMTkiLCJuYmYiOjE2MDI3MDUwMTYsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIiwiaHR0cHM6Ly9pZGVudGl0eS5mb3VuZGF0aW9uLy53ZWxsLWtub3duL2NvbnRleHRzL2RpZC1jb25maWd1cmF0aW9uLXYwLjAuanNvbmxkIl0sImlzc3VlciI6ImRpZDppb246RWlCN0otazBmM0ltRy16eUZVRlJtVjRnYU9zZGZqeTNrZGRKX2Q5T092N3NoQT8taW9uLWluaXRpYWwtc3RhdGU9ZXlKa1pXeDBZVjlvWVhOb0lqb2lSV2xDWmxCVVdqRlliVFI0VG1NNU5uWlFjbmR2V1RBMWJreGZORFZaVUcxNWIxVm5iM2xTYWtGTFFVMWtaeUlzSW5KbFkyOTJaWEo1WDJOdmJXMXBkRzFsYm5RaU9pSkZhVUZGVGswMGNFNUxZa1o2VTJZeFJrMVlSRTVzWVRCVk5sRkNWRzFyWTFCSE9FMXdUR05SWkZaUFRrdG5JbjAuZXlKMWNHUmhkR1ZmWTI5dGJXbDBiV1Z1ZENJNklrVnBRaTFEVFRVMVNGOTJOMHhhUVc5Q1dFazJPWFpFY25BMVgyZDBYMnB0UXpKdWJsaENMVlU1VGxSeFYyY2lMQ0p3WVhSamFHVnpJanBiZXlKaFkzUnBiMjRpT2lKeVpYQnNZV05sSWl3aVpHOWpkVzFsYm5RaU9uc2ljSFZpYkdsalgydGxlWE1pT2x0N0ltbGtJam9pYzJsblh6TTNNemRtTWpJMklpd2lkSGx3WlNJNklrVmpaSE5oVTJWamNESTFObXN4Vm1WeWFXWnBZMkYwYVc5dVMyVjVNakF4T1NJc0ltcDNheUk2ZXlKcmRIa2lPaUpGUXlJc0ltTnlkaUk2SW5ObFkzQXlOVFpyTVNJc0luZ2lPaUpzY2s1UFkxOHliM1JqTTNaVWVFbHVaVEJsY2toVVV5MUdiR0Y0T1VKclEyRldla0pMVkZWZloySm5JaXdpZVNJNkluVlVORmQ0Ym5sUllWcDNMVFI2VGxoQmRUTTBkbGxxU25aM1pGaHVaREo0V2pkeFZtVmhhbWQwYzBVaWZTd2ljSFZ5Y0c5elpTSTZXeUpoZFhSb0lpd2laMlZ1WlhKaGJDSmRmVjE5ZlYxOSIsImlzc3VhbmNlRGF0ZSI6IjIwMjAtMTAtMTRUMTk6NTA6MTYuNzEwWiIsInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLCJEb21haW5MaW5rYWdlQ3JlZGVudGlhbCJdLCJjcmVkZW50aWFsU3ViamVjdCI6eyJpZCI6ImRpZDppb246RWlCN0otazBmM0ltRy16eUZVRlJtVjRnYU9zZGZqeTNrZGRKX2Q5T092N3NoQT8taW9uLWluaXRpYWwtc3RhdGU9ZXlKa1pXeDBZVjlvWVhOb0lqb2lSV2xDWmxCVVdqRlliVFI0VG1NNU5uWlFjbmR2V1RBMWJreGZORFZaVUcxNWIxVm5iM2xTYWtGTFFVMWtaeUlzSW5KbFkyOTJaWEo1WDJOdmJXMXBkRzFsYm5RaU9pSkZhVUZGVGswMGNFNUxZa1o2VTJZeFJrMVlSRTVzWVRCVk5sRkNWRzFyWTFCSE9FMXdUR05SWkZaUFRrdG5JbjAuZXlKMWNHUmhkR1ZmWTI5dGJXbDBiV1Z1ZENJNklrVnBRaTFEVFRVMVNGOTJOMHhhUVc5Q1dFazJPWFpFY25BMVgyZDBYMnB0UXpKdWJsaENMVlU1VGxSeFYyY2lMQ0p3WVhSamFHVnpJanBiZXlKaFkzUnBiMjRpT2lKeVpYQnNZV05sSWl3aVpHOWpkVzFsYm5RaU9uc2ljSFZpYkdsalgydGxlWE1pT2x0N0ltbGtJam9pYzJsblh6TTNNemRtTWpJMklpd2lkSGx3WlNJNklrVmpaSE5oVTJWamNESTFObXN4Vm1WeWFXWnBZMkYwYVc5dVMyVjVNakF4T1NJc0ltcDNheUk2ZXlKcmRIa2lPaUpGUXlJc0ltTnlkaUk2SW5ObFkzQXlOVFpyTVNJc0luZ2lPaUpzY2s1UFkxOHliM1JqTTNaVWVFbHVaVEJsY2toVVV5MUdiR0Y0T1VKclEyRldla0pMVkZWZloySm5JaXdpZVNJNkluVlVORmQ0Ym5sUllWcDNMVFI2VGxoQmRUTTBkbGxxU25aM1pGaHVaREo0V2pkeFZtVmhhbWQwYzBVaWZTd2ljSFZ5Y0c5elpTSTZXeUpoZFhSb0lpd2laMlZ1WlhKaGJDSmRmVjE5ZlYxOSIsIm9yaWdpbiI6Ind3dy5nb29nbGUuY29tIn19fQ.xBd3Q-vka4bkaVBtUjLbimzh1HpzqxWxMaF8eD9l_A56N-81FmOXR-VuEdYmqnTE_TxRwFlDXtaWm_QOXgQXrw"
-    ]
-}"""
-        val response = serializer.parse(DnsBindingResponse.serializer(), docJwt)
-        val domainLinkageCredentialJwt = response.linked_dids.first()
-        runBlocking {
-            val domainLinkageCredential = dnsBindingService.deserializeConfigDocument(domainLinkageCredentialJwt)
-            dnsBindingService.validateWellKnownConfigDocument(domainLinkageCredential)
-        }
-    }*/
-
-}
